+++ source/beamline/AMDetectorTriggerSource.h
@@ -45,16 +45,15 @@ public:
 
 public slots:
 	void requestSetDwellTime(double dwellSeconds);
-    void requestSetDarkCurrentCorrectionTime(int timeSeconds);
-//    AMAction3* doingDarkCurrentCorrection(int newTime);
+    void requestSetDarkCurrentCorrectionTime(double timeSeconds);
 
 	void setSucceeded();
 	void setFailed();
 
 signals:
-	void setDwellTime(double dwellSeconds);
-    void setDarkCurrentCorrectionTime(int timeSeconds);
-    void darkCurrentTimeChanged(int timeSeconds);
+    void setDwellTime(double dwellSeconds);
+    void setDarkCurrentCorrectionTime(double timeSeconds);
+    void darkCurrentTimeChanged(double timeSeconds);
 
 	void succeeded();
 	void failed();
